Move navigation to a separate folder
