+++ src/components/atoms/Header/Header.module.scss
@@ -1,0 +1,19 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-end;
+  background-color: var(--grey);
+}
+
+.welcomeMessage {
+  display: block;
+  width: 85%;
+  margin-bottom: 30px;
+  font-size: 25px;
+  font-weight: bold;
+
+  .firstName {
+    font-weight: normal;
+  }
+}
