+++ src/components/organisms/Navigation/Navigation.module.scss
@@ -1,0 +1,35 @@
+.navigationWrapper {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+  grid-column: 1/1;
+  grid-row: 1/3;
+}
+
+.logo {
+  height: 170px;
+  background-color: white;
+  padding: 50px 10px 50px 50px;
+
+  h2 {
+    color: #005de6;
+  }
+
+  p {
+    margin-top: 10px;
+  }
+}
+
+.navbarItem {
+  p {
+    font-size: 20px;
+  }
+
+  padding: 15px 0 15px 20px;
+  background-color: rgb(54, 205, 247, 0.15);
+  color: var(--blue);
+  width: 90%;
+  border-top-right-radius: 30px;
+  border-bottom-right-radius: 30px;
+}
