+++ src/components/views/Dashboard/Dashboard.module.scss
@@ -43,7 +43,7 @@
   border-bottom-right-radius: 30px;
 }
 
-.dashboard {
+.tableWrapper {
   display: flex;
   justify-content: center;
   align-items: start;
@@ -97,22 +97,5 @@
   color: var(--blue);
 }
 
-.welcomeMessage {
-  display: block;
-  font-size: 25px;
-  font-weight: bold;
-  width: 85%;
-  margin-bottom: 30px;
 
-  .firstName {
-    font-weight: normal;
-  }
-}
 
-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-end;
-  background-color: var(--grey);
-}
